+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.1.0 built on 2019-12-08
+
+ - Fix broken set_block_offsets, block_offsets
+ - Add forgotten tell_compressed
+
 # Version 1.0.0 built on 2019-12-08
 
  - First hopefully stable version